containers/pipefeeder/requirements.txt
bs4
lxml
requests
yt-dlp
flask
flask_sqlalchemy
tqdm
uwsgi
python-dotenv
